Project 109
UCR is a simple, fast and scalable protocol to accelerate the mining of voluntary carbon & water credits from green projects. Universal Carbon Registry was established on 11 July 2021 and is the First and Largest Voluntary Carbon Credits Registry in India. UCR is also the First Registry and Standard to issue Water Offset Credits (RoUs) in the world based on actions taken for capturing/recycling/reusing unutilized sources of water that is socially and culturally equitable, environmentally sustainable and economically beneficial, achieved through site and catchment based actions independent of water quality parameters.